utils: move Git and DVC utilities to packages in linters/versioncontrol, extract mockable exec logic from DVC utils to utils/exec
