+++ linters/versioncontrol/git/git_test.go
@@ -7,7 +7,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/bvobart/mllint/utils/git"
+
+	"github.com/bvobart/mllint/linters/versioncontrol/git"
 )
 
 func TestDetect(t *testing.T) {
